# docs/openapi.yml
openapi: "3.1.0"
info:
  title: "TicketRushServer API"
  description: "TicketRushServer API"
  version: "1.0.2"
servers:
  - url: "https://TicketRushServer"
paths:
...
  /v1/organization:
    get:
      summary: "GET v1/organization"
      operationId: "getInfo"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/FullOrganizationInfo"
    patch:
      summary: "PATCH v1/organization"
      operationId: "updateInfo"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganizationInfoPayload"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationResult"
  /v1/organization/avatar:
    put:
      summary: "PUT v1/organization/avatar"
      operationId: "updateAvatar"
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: "object"
              required:
                - "file"
              properties:
                file:
                  type: "string"
                  format: "binary"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationResult"
  /v1/organization/banner:
    put:
      summary: "PUT v1/organization/banner"
      operationId: "updateBanner"
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: "object"
              required:
                - "file"
              properties:
                file:
                  type: "string"
                  format: "binary"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationResult"
  /v1/organization/events:
    get:
      summary: "GET v1/organization/events"
      operationId: "getEvents"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/BasicEventInfo"
    post:
      summary: "POST v1/organization/events"
      operationId: "createEvent"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventPayload"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationResult"
  /v1/organization/events/{eventId}:
    get:
      summary: "GET v1/organization/events/{eventId}"
      operationId: "getEventDetails"
      parameters:
        - name: "eventId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/FullEventInfo"
    patch:
      summary: "PATCH v1/organization/events/{eventId}"
      operationId: "updateEvent"
      parameters:
        - name: "eventId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventPayload"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationResult"
    delete:
      summary: "DELETE v1/organization/events/{eventId}"
      operationId: "deleteEvent"
      parameters:
        - name: "eventId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationResult"
  /v1/organization/events/{eventId}/banner:
    put:
      summary: "PUT v1/organization/events/{eventId}/banner"
      operationId: "updateEventBanner"
      parameters:
        - name: "eventId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: "object"
              required:
                - "file"
              properties:
                file:
                  type: "string"
                  format: "binary"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationResult"
  /v1/organization/events/{eventId}/publish:
    post:
      summary: "POST v1/organization/events/{eventId}/publish"
      operationId: "publishEvent"
      parameters:
        - name: "eventId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationResult"
  /v1/organization/events/{eventId}/sales-rounds:
    post:
      summary: "POST v1/organization/events/{eventId}/sales-rounds"
      operationId: "addSalesRound"
      parameters:
        - name: "eventId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddSalesRoundPayload"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationResult"
  /v1/organization/events/{eventId}/sales-rounds/{roundId}:
    patch:
      summary: "PATCH v1/organization/events/{eventId}/sales-rounds/{roundId}"
      operationId: "updateSalesRound"
      parameters:
        - name: "eventId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "roundId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSalesRoundData"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationResult"
    delete:
      summary: "DELETE v1/organization/events/{eventId}/sales-rounds/{roundId}"
      operationId: "deleteSalesRound"
      parameters:
        - name: "eventId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "roundId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationResult"
  /v1/organization/events/{eventId}/seat-zones:
    post:
      summary: "POST v1/organization/events/{eventId}/seat-zones"
      operationId: "createSeatZone"
      parameters:
        - name: "eventId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSeatZonePayload"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationResult"
  /v1/organization/events/{eventId}/seat-zones/{zoneId}:
    delete:
      summary: "DELETE v1/organization/events/{eventId}/seat-zones/{zoneId}"
      operationId: "deleteSeatZone"
      parameters:
        - name: "eventId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "zoneId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationResult"
  /v1/organization/events/{eventId}/staff:
    post:
      summary: "POST v1/organization/events/{eventId}/staff"
      operationId: "addEventStaffAccount"
      parameters:
        - name: "eventId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddEventStaffPayload"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationResult"
  /v1/organization/events/{eventId}/staff/{staffId}:
    delete:
      summary: "DELETE v1/organization/events/{eventId}/staff/{staffId}"
      operationId: "deleteEventStaffAccount"
      parameters:
        - name: "eventId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "staffId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationResult"
  /v1/organization/events/{eventId}/ticket-classes:
    post:
      summary: "POST v1/organization/events/{eventId}/ticket-classes"
      operationId: "createTicketClass"
      parameters:
        - name: "eventId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTicketClassPayload"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationResult"
  /v1/organization/events/{eventId}/ticket-classes/{ticketClassId}:
    delete:
      summary: "DELETE v1/organization/events/{eventId}/ticket-classes/{ticketClassId}"
      operationId: "deleteTicketClass"
      parameters:
        - name: "eventId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
        - name: "ticketClassId"
          in: "path"
          required: true
          schema:
            type: "integer"
            format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationResult"
...
  /v1/user:
    get:
      summary: "GET v1/user"
      operationId: "getInfo"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/FullUserInfo"
    patch:
      summary: "PATCH v1/user"
      operationId: "updateInfo"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserInfoPayload"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationResult"
  /v1/user/avatar:
    put:
      summary: "PUT v1/user/avatar"
      operationId: "updateAvatar"
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: "object"
              required:
                - "file"
              properties:
                file:
                  type: "string"
                  format: "binary"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OperationResult"
...
components:
  schemas:
...
    OperationResult:
      type: "object"
      properties:
        success:
          type: "boolean"
          nullable: true
        message:
          type: "string"
          nullable: true
        resourceId:
          type: "integer"
          format: "int64"
          nullable: true
...
    Instant:
      type: "object"
      properties: { }
    BasicEventInfo:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          nullable: true
        name:
          type: "string"
          nullable: true
        bannerUrl:
          type: "string"
          nullable: true
        dateTime:
          nullable: true
          $ref: "#/components/schemas/Instant"
        venue:
          type: "string"
          nullable: true
...
    FullOrganizationInfo:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          nullable: true
        name:
          type: "string"
          nullable: true
        email:
          type: "string"
          nullable: true
        createdAt:
          nullable: true
          $ref: "#/components/schemas/Instant"
        updatedAt:
          nullable: true
          $ref: "#/components/schemas/Instant"
        verified:
          type: "boolean"
          nullable: true
        description:
          type: "string"
          nullable: true
        aliasName:
          type: "string"
          nullable: true
        avatarUrl:
          type: "string"
          nullable: true
        bannerUrl:
          type: "string"
          nullable: true
        websiteUrl:
          type: "string"
          nullable: true
    UpdateOrganizationInfoPayload:
      type: "object"
      properties:
        description:
          type: "string"
          nullable: true
        aliasName:
          type: "string"
          nullable: true
        websiteUrl:
          type: "string"
          nullable: true
    CreateEventPayload:
      type: "object"
      properties:
        name:
          type: "string"
          nullable: true
        description:
          type: "string"
          nullable: true
        isOnlineEvent:
          type: "boolean"
          nullable: true
        venue:
          type: "string"
          nullable: true
        address:
          type: "string"
          nullable: true
        dateTime:
          nullable: true
          $ref: "#/components/schemas/Instant"
    FullEventInfo:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          nullable: true
        name:
          type: "string"
          nullable: true
        description:
          type: "string"
          nullable: true
        published:
          type: "boolean"
          nullable: true
        isOnlineEvent:
          type: "boolean"
          nullable: true
        venue:
          type: "string"
          nullable: true
        address:
          type: "string"
          nullable: true
        dateTime:
          nullable: true
          $ref: "#/components/schemas/Instant"
        bannerUrl:
          type: "string"
          nullable: true
        createdAt:
          nullable: true
          $ref: "#/components/schemas/Instant"
        updatedAt:
          nullable: true
          $ref: "#/components/schemas/Instant"
    UpdateEventPayload:
      type: "object"
      properties:
        name:
          type: "string"
          nullable: true
        description:
          type: "string"
          nullable: true
        isOnlineEvent:
          type: "boolean"
          nullable: true
        venue:
          type: "string"
          nullable: true
        address:
          type: "string"
          nullable: true
        dateTime:
          nullable: true
          $ref: "#/components/schemas/Instant"
    AddSalesRoundPayload:
      type: "object"
      properties:
        name:
          type: "string"
          nullable: true
        startTime:
          nullable: true
          $ref: "#/components/schemas/Instant"
        endTime:
          nullable: true
          $ref: "#/components/schemas/Instant"
        maxTicketsPerPurchase:
          type: "integer"
          format: "int32"
          nullable: true
    UpdateSalesRoundData:
      type: "object"
      properties:
        name:
          type: "string"
          nullable: true
        startTime:
          nullable: true
          $ref: "#/components/schemas/Instant"
        endTime:
          nullable: true
          $ref: "#/components/schemas/Instant"
        maxTicketsPerPurchase:
          type: "integer"
          format: "int32"
          nullable: true
    SeatView:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          nullable: true
        index:
          type: "integer"
          format: "int32"
          nullable: true
        number:
          type: "integer"
          format: "int32"
          nullable: true
        availability:
          type: "string"
          nullable: true
          enum:
            - "AVAILABLE"
            - "SOLD"
            - "HELD_BY_ME"
            - "HELD"
    SeatRowView:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          nullable: true
        index:
          type: "integer"
          format: "int32"
          nullable: true
        label:
          type: "string"
          nullable: true
        seats:
          type: "array"
          nullable: true
          items:
            $ref: "#/components/schemas/SeatView"
    CreateSeatZonePayload:
      type: "object"
      properties:
        name:
          type: "string"
          nullable: true
        positionX:
          type: "integer"
          format: "int32"
          nullable: true
        positionY:
          type: "integer"
          format: "int32"
          nullable: true
        rows:
          type: "array"
          nullable: true
          items:
            $ref: "#/components/schemas/SeatRowView"
    AddEventStaffPayload:
      type: "object"
      properties:
        name:
          type: "string"
          nullable: true
        email:
          type: "string"
          nullable: true
        password:
          type: "string"
          nullable: true
    CreateTicketClassPayload:
      type: "object"
      properties:
        name:
          type: "string"
          nullable: true
        description:
          type: "string"
          nullable: true
        price:
          type: "integer"
          format: "int64"
          nullable: true
        salesRoundId:
          type: "integer"
          format: "int64"
          nullable: true
        seatZoneId:
          type: "integer"
          format: "int64"
          nullable: true
...
    FullUserInfo:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          nullable: true
        name:
          type: "string"
          nullable: true
        email:
          type: "string"
          nullable: true
        avatarUrl:
          type: "string"
          nullable: true
        birthDate:
          type: "string"
          format: "date-time"
          nullable: true
        gender:
          type: "string"
          nullable: true
        phoneNumber:
          type: "string"
          nullable: true
        addressLine:
          type: "string"
          nullable: true
        createdAt:
          nullable: true
          $ref: "#/components/schemas/Instant"
        updatedAt:
          nullable: true
          $ref: "#/components/schemas/Instant"
    UpdateUserInfoPayload:
      type: "object"
      properties:
        birthDate:
          type: "string"
          format: "date-time"
          nullable: true
        gender:
          type: "string"
          nullable: true
          enum:
            - "MALE"
            - "FEMALE"
            - "OTHER"
        phoneNumber:
          type: "string"
          nullable: true
        addressLine:
          type: "string"
          nullable: true
...